+++ Assets/ActiveRay.cs
@@ -1,0 +1,20 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.XR.Interaction.Toolkit;
+using UnityEngine.InputSystem;
+
+public class ActiveRay : MonoBehaviour
+{
+    //public GameObject RightTp; //광선오브젝트 
+
+    //public InputActionProperty RightActivate; //오른쪽 트리거 버튼 입력을 받음
+
+    //// Update is called once per frame
+    //void Update()
+    //{
+    //    RightTp.SetActive(RightActivate.action.ReadValue<float>() > 0.1f);
+    //    //오른쪽 트리거의 버튼 값이 0.1보다 크면 오른쪽 텔레포트 광선 활성
+        
+    //}
+}
