--- Assets/Font/UnityFont.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5b98297789a87304cbb46ece6191c20d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
