--- Assets/Font/UnityFont/Jaemin3-Regular SDF.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cfef4f5ab75a0c54cb94a6be18384e1d
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
