--- Assets/Font/UnityFont/MYArirang SDF.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 180b4ac0a00047640a94e28a92226502
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
