--- Assets/Font/UnityFont/MYHaemalgeunSangsang SDF.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 328e20cccb261064094c8f5340ce0102
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
