--- Assets/Font/UnityFont/MYYeongnamnu SDF.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 336eca60645a89a4597eecdc363906f8
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
