+++ Assets/HandTracking.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e457a9b6480d84743a556b2c7b79a8fc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
