+++ Assets/Materials/TeleportPlane Green.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e7b1652c1e06a204b8f842c377267b8f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
