--- Assets/NatureManufacture Assets/Forest Environment Dynamic Nature/Demo Scenes/Forest Demo Scene/NavMesh.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2a25a1037b742a24286518059de1b9cd
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
