--- Assets/New Terrain.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 28be86d48c47dc9429d4d4e34a68eddd
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 15600000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
