--- Assets/Oculus Hands/Materials/Teleport_Materials.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 16c828ee0fddf64459a0ac12910c0c68
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
