--- Assets/Oculus Hands/Materials/White_Materials.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 189c213b5e5b06247bcfce7895f5eae4
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
