--- Assets/Player.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-<<<<<<<< HEAD:Assets/Polybrush Data.meta
-guid: d72563dc13f18f541823e8be9e7ff2e8
-========
-guid: 772f10744beeab649ae73586c5b1a4a9
->>>>>>>> main:Assets/Player.meta
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
