--- Assets/Player/HandTracking.cs
@@ -1,21 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine.InputSystem;
-using UnityEngine;
-
-public class HandTracking : MonoBehaviour
-{
-    public InputActionProperty pinchAnimaionAction;
-    public InputActionProperty gripAnimtionAction;
-
-    public Animator HandAnimator;
-
-    void Update()
-    {
-        float triggerValue = pinchAnimaionAction.action.ReadValue<float>();
-        HandAnimator.SetFloat("Trigger", triggerValue);
-        float gripValue = gripAnimtionAction.action.ReadValue<float>();
-        HandAnimator.SetFloat("Grip", gripValue);
-        
-    }
-}
