--- Assets/Scenes/TigerTestScene.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c9edb67852b9b7a42b540f771cea7705
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
