--- Assets/Scenes/TigerTestScene/NavMesh.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e1b6f99c060a394459142ca8a0de806f
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 23800000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
