--- Assets/Scripts/Tiger/Test.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7be02fe8c50f39d4d90897869d464985
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
