--- Assets/Scripts/Tiger/Test/GameMgr.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 411b970dda150eb4b810e4d8de7ecb6d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
