--- Assets/Scripts/Tiger/Test/PlayerTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b80986cff5a813a45807572db77910d0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
