--- Assets/Scripts/Tiger/Test/TestLevelChanger.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8cad16076c0b94545a677979f2684e14
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
