+++ Assets/Scripts/Tiger/Tiger.cs
@@ -1,12 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.AI;
 using UnityEngine.UIElements;
 
 public class Tiger : MonoBehaviour
 {
     // 필요 변수 (호랑이 속도, 움직임 상태, 
     private Vector3 spawnPos;
-
     [Header("TigerStateInfo")]
     public float tigerWalkSpeed;
     public float tigerRunSpeed;
@@ -15,18 +15,8 @@ public class Tiger : MonoBehaviour
 
     private Animator animator;
 
-    private NavMeshAgent meshAgent;
-
     [HideInInspector]
     public Transform playerT;
-
-    // 호랑이 자동 움직임 변수
-    private float updateInterval = 10f;
-    private float timeSinceLastUpdate;
-
-    // 범위 스캔 체크 변수
-    public int checkRange = 0;
-    private int updateCheckRange = 0;
 
     public enum TState
     {
@@ -44,14 +34,11 @@ public class Tiger : MonoBehaviour
         playerT = transform;
         tigerState = TState.Idle;
         animator = GetComponent<Animator>();
-        meshAgent = GetComponent<NavMeshAgent>();
-        timeSinceLastUpdate = updateInterval;
     }
 
     // Update is called once per frame
     void Update()
     {
-        
         // 플레이어 좌표를 향해 일정 속도로 달려오기. -> 플레이어와 일정범위 안으로 가까워지면 점프 공격
         // 곶감에 당했을 때 (도망)
 
@@ -61,107 +48,26 @@ public class Tiger : MonoBehaviour
         {
             case TState.Idle:
                 Debug.Log("Idle");
-                GetRandomMove();
+                ScanPlayer();
                 TigerMove();
                 break;
             case TState.Run:
                 Debug.Log("Run");
-                GetRandomMove();
+                ScanPlayer();
                 TigerRun();
                 break;
             case TState.Attack:
                 Debug.Log("Attack");
-                //ScanPlayer();
+                ScanPlayer();
                 TigerAttack();
                 break;
         }
         // 이동 테스트
         
     }
-
-    public void OnCheckScanRange()
-    {
-        var player = GameMgr.Instance.PlayerInit();
-        if (updateCheckRange != checkRange)
-        {
-            switch (checkRange)
-            {
-                case 0:
-                    break;
-                case 1:
-                    player.SetUpBoundaryLevel();
-                    break;
-                case 2:
-                    player.SetDownBoundaryLevel();
-                    break;
-            }
-
-            Debug.Log("CheckRange : " + checkRange);
-            updateCheckRange = checkRange;
-        }
-        
-        
-    }
-
-    // 3초마다 새로운 랜덤 위치를 찾아 이동하는 함수
-    public void GetRandomMove()
-    {
-        timeSinceLastUpdate += Time.deltaTime;
-        OnCheckScanRange();
-        if (timeSinceLastUpdate >= updateInterval)
-        {
-            if(GameMgr.Instance.PlayerInit().GetBoundaryLevel() <= 0)
-            {
-                Vector3 randPos = GetRandomPosition();
-                meshAgent.SetDestination(randPos);
-                timeSinceLastUpdate = 0f;
-            }
-            else 
-            {
-                meshAgent.SetDestination(playerT.position);
-                timeSinceLastUpdate = 0f;
-            }
-        }
-    }
-    
-    public Vector3 GetRandomPosition()
-    {
-        Vector3 randDirect = Random.insideUnitSphere * 20f;
-        randDirect += transform.position;
-
-        NavMeshHit hit;
-        if(NavMesh.SamplePosition(randDirect, out hit, 20f, NavMesh.AllAreas))
-        {
-            return hit.position;
-        }
-        else
-        {
-            return transform.position;
-        }
-        
-    }
-
     // 상태 변화 함수(정지, 공격, 정찰)
-    public void TigerStateChanger()
+    public void TigerStateChange(TState state)
     {
-        int level = GameMgr.Instance.PlayerInit().GetBoundaryLevel();
-        Debug.Log("BoundaryLevel = " + level);
-        switch (level)
-        {
-            case 0:
-                tigerState = TState.Idle;
-                break;
-            case 1:
-                tigerState = TState.Idle;
-                break;
-            case 2:
-                tigerState = TState.Run;
-                break;
-            case 3:
-                tigerState = TState.Attack;
-                break;
-        }
-        GetRandomMove();
     }
     // 움직임 함수
     private void TigerMove()
@@ -169,20 +75,18 @@ public class Tiger : MonoBehaviour
         // 걷기 애니메이션 재생
 
         //걷기 코드
-        //float speed = tigerWalkSpeed * Time.deltaTime;
-        //transform.Translate(Vector3.forward * speed);
-        meshAgent.speed = tigerWalkSpeed;
+        float speed = tigerWalkSpeed * Time.deltaTime;
+        transform.Translate(Vector3.forward * speed);
         animator.SetBool("ScanTigerL", false);
         animator.SetBool("ScanTigerS", true);
 
     }
     private void TigerRun()
     {
-
+        
         //뛰기 코드
-        //float speed = tigerRunSpeed * Time.deltaTime;
-        //transform.Translate(Vector3.forward * speed);
-        meshAgent.speed = tigerRunSpeed;
+        float speed = tigerRunSpeed * Time.deltaTime;
+        transform.Translate(Vector3.forward * speed);
         animator.SetBool("ScanTigerS", false);
         animator.SetBool("ScanTigerL", true);
 
@@ -194,21 +98,16 @@ public class Tiger : MonoBehaviour
         animator.SetBool("AttackTiger", true);
     }
 
-    // 플레이어 감지 함수(사용 안함)
+    // 플레이어 감지 함수
     private void ScanPlayer()
     {
         // 충돌 판정 -> 태그 플레이어
         // if((transform.position)
         // 플레이어 방향 바라보기
-
-        // 위치 이동
-        //vector3 dir = playert.position - transform.position;
-        //dir.y = 0f;
-        //quaternion rot = quaternion.lookrotation(dir.normalized);
-        //transform.rotation = rot;
-        
-        // navMesh 사용 이동
-        meshAgent.SetDestination(playerT.position);
+        Vector3 dir = playerT.position - transform.position;
+        dir.y = 0f;
+        Quaternion rot = Quaternion.LookRotation(dir.normalized);
+        transform.rotation = rot;
 
         // 플레이어에게 달려가기 방향만 바라보고 애니메이션으로 달리기.
