+++ Assets/Scripts/Tiger/TigerAttack.cs
@@ -11,10 +11,8 @@ public class TigerAttack : MonoBehaviour
         Debug.Log("플레이어 감지");
         if (other.CompareTag("Player"))
         {
-            // 슬로우모션
-            // GameMgr.Instance.SlowMotion(0.5f);
-            GameMgr.Instance.PlayerInit().SetBoundaryLevel(3);
-            GetTiger.TigerStateChanger();
+            GetTiger.tigerState = Tiger.TState.Attack;
+
         }
     }
     private void OnTriggerStay(Collider other)
@@ -22,8 +20,6 @@ public class TigerAttack : MonoBehaviour
         if (other.CompareTag("Player"))
         {
             Debug.Log("플레이어 감지중");
-            GetTiger.TigerStateChanger();
-            
             GetTiger.playerT = other.transform;
         }
     }
