+++ Assets/Scripts/Tiger/TigerScanL.cs
@@ -11,26 +11,22 @@ public class TigerScanL : MonoBehaviour
         Debug.Log("플레이어 감지");
         if (other.CompareTag("Player"))
         {
-            GetTiger.checkRange = 1;
-            GetTiger.TigerStateChanger();
+            GetTiger.tigerState = Tiger.TState.Run;
+
         }
     }
     private void OnTriggerStay(Collider other)
     {
-
         if (other.CompareTag("Player"))
         {
             Debug.Log("플레이어 감지중");
             GetTiger.playerT = other.transform;
-            GetTiger.TigerStateChanger();
         }
     }
 
     private void OnTriggerExit(Collider other)
     {
         Debug.Log("플레이어 감지 종료");
-        GetTiger.checkRange = 0;
-        GameMgr.Instance.PlayerInit().SetDownBoundaryLevel();
         GetTiger.tigerState = Tiger.TState.Idle;
     }
