+++ Assets/XRI/Settings/Resources/InteractionLayerSettings.asset
@@ -14,7 +14,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   m_LayerNames:
   - Default
-  - Teleport
+  - 
   - 
   - 
   -
