+++ ProjectSettings/NavMeshAreas.asset
@@ -69,9 +69,9 @@ NavMeshProjectSettings:
     cost: 1
   - name: 
     cost: 1
-  m_LastAgentTypeID: -334000983
+  m_LastAgentTypeID: -887442657
   m_Settings:
-  - serializedVersion: 3
+  - serializedVersion: 2
     agentTypeID: 0
     agentRadius: 0.5
     agentHeight: 2
@@ -84,29 +84,8 @@ NavMeshProjectSettings:
     cellSize: 0.16666667
     manualTileSize: 0
     tileSize: 256
-    buildHeightMesh: 0
-    maxJobWorkers: 0
-    preserveTilesOutsideBounds: 0
-    debug:
-      m_Flags: 0
-  - serializedVersion: 3
-    agentTypeID: -334000983
-    agentRadius: 1
-    agentHeight: 1
-    agentSlope: 50
-    agentClimb: 0.6
-    ledgeDropHeight: 0
-    maxJumpAcrossDistance: 2
-    minRegionArea: 2
-    manualCellSize: 0
-    cellSize: 0.16666667
-    manualTileSize: 0
-    tileSize: 256
-    buildHeightMesh: 0
-    maxJobWorkers: 0
-    preserveTilesOutsideBounds: 0
+    accuratePlacement: 0
     debug:
       m_Flags: 0
   m_SettingNames:
   - Humanoid
-  - Tiger
