+++ ProjectSettings/TagManager.asset
@@ -9,7 +9,7 @@ TagManager:
   - Default
   - TransparentFX
   - Ignore Raycast
-  - Teleport
+  - Ground
   - Water
   - UI
   -
